+++ example/lib/main.dart
@@ -5,6 +5,7 @@ import 'package:intl/intl.dart' as intl;
 import 'sample10.dart';
 import 'sample12.dart';
 import 'sample13.dart';
+import 'sample14.dart';
 
 void main() => runApp(MyApp());
 
@@ -143,7 +144,15 @@ class MyHomePage extends StatelessWidget {
               subtitle: Text("Dynamic date range"),
               onTap: () => Navigator.of(context)
                   .push(MaterialPageRoute(builder: (context) => Sample13())),
-            ),          ],
+            ),
+            ListTile(
+              title: Text("Sample 14"),
+              subtitle: Text(
+                  "Sample with updatePositionOnTap & format indicator value"),
+              onTap: () => Navigator.of(context)
+                  .push(MaterialPageRoute(builder: (context) => Sample14())),
+            ),
+          ],
         ),
       ),
     );
@@ -350,6 +359,9 @@ Widget sample3(BuildContext context) {
         onDateTimeSelected: (datetime) {
           print("selected datetime: $datetime");
         },
+        onScaleChanged: (scale) {
+          print("Scale: $scale");
+        },
         selectedDate: toDate,
         //this is optional
         footerDateTimeBuilder: (DateTime value, BezierChartScale scaleType) {
@@ -374,7 +386,7 @@ Widget sample3(BuildContext context) {
         config: BezierChartConfig(
           displayDataPointWhenNoValue: false,
           verticalIndicatorStrokeWidth: 3.0,
-          pinchZoom: false,
+          pinchZoom: true,
           physics: ClampingScrollPhysics(),
           verticalIndicatorColor: Colors.black26,
           showVerticalIndicator: true,